more

+++ python/sglang/srt/managers/expert_location.py
@@ -5,7 +5,6 @@ from dataclasses import dataclass
 from typing import List, Optional
 
 import torch
-
 from sglang.srt.configs.model_config import ModelConfig
 from sglang.srt.managers import deepseek_eplb
 from sglang.srt.model_loader import get_model_architecture
@@ -34,8 +33,8 @@ class ExpertLocationMetadata:
         num_logical_experts = model_config_for_expert_location.num_logical_experts
 
         physical_to_logical_map = (
-            torch.arange(0, num_physical_experts).repeat(num_layers, 1)
-            % num_logical_experts
+                torch.arange(0, num_physical_experts).repeat(num_layers, 1)
+                % num_logical_experts
         )
         # Highly inefficient, but we do not care since we will use EPLB distribution logic
         logical_to_all_physical_map = torch.arange(0, num_logical_experts).repeat(
@@ -50,7 +49,7 @@ class ExpertLocationMetadata:
 
     @staticmethod
     def init_by_mapping(
-        server_args: ServerArgs, physical_to_logical_map, logical_to_all_physical_map
+            server_args: ServerArgs, physical_to_logical_map, logical_to_all_physical_map
     ):
         if not isinstance(physical_to_logical_map, torch.Tensor):
             physical_to_logical_map = torch.tensor(physical_to_logical_map)
@@ -107,8 +106,8 @@ class ExpertLocationMetadata:
         )
 
         num_physical_experts = (
-            model_config_for_expert_location.num_logical_experts
-            + server_args.ep_num_redundant_experts
+                model_config_for_expert_location.num_logical_experts
+                + server_args.ep_num_redundant_experts
         )
         # TODO consider case when DP attention is disabled and DP > 1
         world_size = server_args.tp_size
@@ -156,7 +155,7 @@ class ExpertLocationMetadata:
         return global_physical_expert_index % self.num_local_physical_experts
 
     def logical_to_all_physical(
-        self, layer_id: int, logical_expert_id: int
+            self, layer_id: int, logical_expert_id: int
     ) -> List[int]:
         return self.logical_to_all_physical_raw(
             self.logical_to_all_physical_map, layer_id, logical_expert_id
@@ -164,7 +163,7 @@ class ExpertLocationMetadata:
 
     @staticmethod
     def logical_to_all_physical_raw(
-        logical_to_all_physical_map, layer_id: int, logical_expert_id: int
+            logical_to_all_physical_map, layer_id: int, logical_expert_id: int
     ) -> List[int]:
         return [
             physical_expert_id
@@ -184,8 +183,8 @@ class ExpertLocationMetadata:
 
 
 def _compute_logical_to_rank_dispatch_physical_map(
-    logical_to_all_physical_map: torch.Tensor,
-    num_gpus: int,
+        logical_to_all_physical_map: torch.Tensor,
+        num_gpus: int,
 ):
     # TODO maybe improve this algorithm (e.g. ensure it is really balanced)
     # This is rarely called, so we use for loops for maximum clarity
@@ -194,7 +193,8 @@ def _compute_logical_to_rank_dispatch_physical_map(
 
     num_layers, num_logical_experts, _ = logical_to_all_physical_map.shape
     logical_to_rank_dispatch_physical_map = torch.zeros(
-        (num_gpus, num_layers, num_logical_experts)
+        (num_gpus, num_layers, num_logical_experts),
+        dtype=logical_to_all_physical_map.dtype,
     )
 
     for layer_id in range(num_layers):
